Have enemy's increase in size linearly with respect to the player
Enemies were being spawned with a radius up to twice the size of the player
which meant that the size was increasing as the square of the player's change
in size.

Using the square root of the player's radius as the non-constant factor in
determining enemy size ensures that there is not a sudden surge in enemy size
when the player approaches the 'end-game'.

A constant factor of 12.0 was found to be relatively 'good' for ensuring that
the player has the oppurtunity to win but is not able to easily win every
game. This factor could be adjusted as part of a difficulty selection, however
most likely further rework of the function would be required.